requirement.txt
flask
flask-cors
requests
tensorflow
numpy
Pillow
